Fixing 64bit revent
Then number of input event files was being written as a size_t but read
as an int by revent. These types have different sizes with 64bit GCC,
causing revent not being able to replay recorded files. This comint
updates revent to use size_t when both reading and writing.

+++ wlauto/external/revent/revent.c
@@ -81,8 +81,8 @@ typedef struct {
 } replay_event_t;
 
 typedef struct {
-    int num_fds;
-    int num_events;
+    size_t num_fds;
+    size_t num_events;
     int *fds;
     replay_event_t *events;
 } replay_buffer_t;
